spef_extractor: remove unused files
Many files in lef_def_parser/ are not used by the spef_extractor.

This clarifies was is indeed used (and also save a little bit disk space).
This increases the divergence with the upstream lef_def_parser, but the
latter hasn't been updated for years and the remaining files (lef and
def parser) are not that complex.