resulting magic files are now checked by the comparison scripts
I added a check for the design to design comparison. This check verifies that the magic resulting files exist and reports that in the pass and fail report.
This check was only added to the compare_regression_design.py (A design choice) because compare_regression_reports.py, as the name indicates, is only concerned with the summary reports.
An assumption was made that since the lef, gds, mag, and spice files exist in the final magic results, then all previous files must've existed, otherwise there would be lvs/drc issues and so the test will report failure as well.

This extra check was added as a result of the incident when we updated a magic version that passed the tests, but it turned out that it had a broken 'lef write' command and so produced no .lef files.