[CI] rework the way testsets are handle
It doesn't make any sense to have files containing only their name saved inside. So, let's relocate the actual test sets into a test_sets directory and adjust the test script to detect whether the provided env var is one of the test sets, if not then it's a design name and the CI should act accordingly.