[opendp] Restore diamond_search_height_ to 400
diamond_search_height_ was originally 400 but has since been changed
to 100. On a design with large macros, the openroad version of opendp
fails but the standalone version works.

The upstream fix to openroad is still being discussed in:

https://github.com/The-OpenROAD-Project/OpenROAD/pull/549

but since it is causing issues with at least one caravel design, restore
the old behaviour until the ultimate fix makes it in.