Speed up extract_antenna_violators.py
extract_antenna_violators.py is doing expensive splits across the
entire file which requires large mallocs and memcpys. It also
reads the entire file in before parsing it.

Replace this with a simpler and faster way of achieving the same
thing. On my test case the time goes from over 12 minutes to 0.62
seconds, about 1200x faster.