Multi-Stage Builds For Tools
Tool images no longer include source code or build artifacts, but they are still runnable. The result images are a lot smaller, but can still be run independently.

This entailed the creation of a common base image for runnable images.

The images still contain tarballs for backwards compatibility with older versions of OpenLane.