Include all PDK lib files when merging
When using asap7, the OpenLane flow fails to complete synthesis. The issue is that many cells which are required to perform technology mapping cannot be found by `abc`, even though they're provided in the PDK. This is because when merging Liberty files, the current script fails to provide all files listed by the `LIB_SYNTH_COMPLETE` environment variable as input. This patch fixes that.