Use LIB_SYNTH instead of LIB_SYNTH_COMPLETE in some OpenROAD scripts
When using asap7, LIB_SYNTH_COMPLETE specifies several files, instead of a single one, like many commands in the flow scripts expect. To work around this, we can specify LIB_SYNTH instead, which, after `prep` in all.tcl is invoked, refers to the output of Liberty file merging, which should contain everything we need to continue in the flow.