sed: use cross-platform form for -i flags
The `-i` argument to sed has differing behaviour under BSD and GNU sed.

On BSD it requires an argument, though it can be '' (i.e. `-i ''`).

Under GNU sed, it needs to immediately follow the argument (i.e.
`-i''`).

Provide a backup file argument and immediately delete it, since this
appears to be the only cross-platform solution.

Signed-off-by: Sean Cross <sean@xobs.io>