Adding the option the use welltap-less PDK (#2002)
Some PDKs don't have welltaps as dedicated cells, typically high voltage technologies with isolation wells. I have slightly modified the scripts such that if no tapcells are defined, it still inserts the decap cells instead of skipping the step altogether.

I am not certain if the `place_endcaps` commands should be placed before or after the `cut_rows` command, but it appears to work for me as it is.